requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
urllib3==2.2.2
numpy==2.2.5
pandas==2.2.3
yfinance==0.2.56
TA-Lib==0.5.1